Completely restructure installation documentation
This work breaks apart the installation page, which was very long. It also
establishes a new organization method for the installation section as a
whole, sorting .rst files into appropriate directories for better navigation
by editors. This also updates the various indexes, toctrees, and navigation
elements for the installation docs.

Resolves #94.